Move GitFindRootFromPath to the Repository category
