+++ src/Cake.Git/Cake.Git.csproj
@@ -65,7 +65,6 @@
     <Compile Include="GitAliases.cs" />
     <Compile Include="GitAliases.Describe.cs" />
     <Compile Include="GitAliases.Diff.cs" />
-    <Compile Include="GitAliases.GitFindRootFromPath.cs" />
     <Compile Include="GitAliases.Init.cs" />
     <Compile Include="GitAliases.Log.cs" />
     <Compile Include="GitAliases.Pull.cs" />
